split process.go
